Using Khan's replacement transitions component for React, instead of the default one, since the default one doesn't work very well.
Default transition component has a bug where if a certain event is not fired it will not complete the animation, which can be triggered by the browser being in a different tab, or any reason it wants really. This is kind of a showstopper for using the default transitions since it lead to removed elements not being removed since the transition never completed.